pyproject.toml
[tool.poetry]
name = "django-peeringdb"
version = "3.4.0"
description = "PeeringDB Django models"
readme = "README.md"
repository = "https://github.com/peeringdb/django-peeringdb"
authors = ["PeeringDB <support@peeringdb.com>"]
license = "Apache-2.0"

classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Framework :: Django :: 3.1",
  "Framework :: Django :: 3.2",
  "Framework :: Django :: 4.0",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Internet",
]

packages = [{ include = "django_peeringdb", from = "src" }]

[tool.poetry.dependencies]
python = "^3.8"
django_countries = ">1"
django_handleref = ">=2"
django_inet = ">=1"
asgiref = ">=3"

[tool.poetry.group.dev.dependencies]
# testing
codecov = ">=2"
coverage = ">=5"
pytest = ">=6"
pytest-django = ">=3.8"
pytest-cov = "*"
tox = ">=3.24"

# linting
black = ">=20"
isort = ">=5.7"
flake8 = ">=3.8"
mypy = ">=0.950"
pre-commit = ">=2.13"
pyupgrade = ">=2.19"
ruff = "^0.1.11"

# docs
markdown = "*"
markdown-include = ">=0.5,<1"
mkdocs = ">=1.2.3"

# ctl
ctl = ">=1"
jinja2 = ">=2"
tmpl = ">=1"
twine = ">=3.3"

[tool.poetry.plugins."markdown.extensions"]
pymdgen = "pymdgen.md:Extension"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3

